// src/interfaces.ts
import React from "react";

export interface TodoInterface {
    id: string;
    text: string;
    isCompleted: boolean;
}

export interface TodoFormInterface {
    todos: TodoInterface[];
    handleTodoCreate: (todo: TodoInterface) => void;
}

export interface TodoListInterface {
    handleTodoUpdate: (
        event: React.ChangeEvent<HTMLInputElement>,
        id: string
    ) => void;
    handleTodoRemove: (id: string) => void;
    handleTodoComplete: (id: string) => void;
    handleTodoBlur: (event: React.ChangeEvent<HTMLInputElement>) => void;
    todos: TodoInterface[];
}

export interface TodoItemInterface {
    handleTodoUpdate: (
        event: React.ChangeEvent<HTMLInputElement>,
        id: string
    ) => void;
    handleTodoRemove: (id: string) => void;
    handleTodoComplete: (id: string) => void;
    handleTodoStatus: (event: React.MouseEvent<HTMLElement>) => void;
    handleTodoBlur: (event: React.ChangeEvent<HTMLInputElement>) => void;
    todo: TodoInterface;
}